Refactor amenity toggle logic into a separate hook
